+++ test/test_qat.py
@@ -7,7 +7,7 @@ import torch
 from torch.nn import Conv2d, BatchNorm2d, ReLU
 from torch.nn._intrinsic.qat import ConvBn2d, ConvBnReLU2d
 from torch.quantization.QConfig import default_qat_qconfig
-import torch.backends.mkldnn as mkldnn
+import torch.backends.mkldnn
 from common_utils import TestCase, run_tests
 from hypothesis import given
 from hypothesis import strategies as st
@@ -59,6 +59,10 @@ class IntrinsicQATModuleTest(TestCase):
             momentum,
             freeze_bn
     ):
+        # **** WARNING: This is used to temporarily disable MKL-DNN convolution due
+        # to a bug: https://github.com/pytorch/pytorch/issues/23825
+        # Once this bug is fixed, this context manager as well as its callsites
+        # should be removed!
         with torch.backends.mkldnn.flags(enabled=False):
             input_channels = input_channels_per_group * groups
             output_channels = output_channels_per_group * groups
